tranger in c_tranger

+++ src/c_tranger.c
@@ -1432,7 +1432,7 @@ PRIVATE json_t *cmd_record_pkey2s(hgobj gobj, const char *cmd, json_t *kw, hgobj
     return msg_iev_build_webix(
         gobj,
         pkey2s?0:-1,
-        json_sprintf("%d pkey2s", json_object_size(pkey2s)),
+        json_sprintf("%d pkey2s", (int)json_object_size(pkey2s)),
         0,
         pkey2s,
         kw  // owned
